scripts/requirements.txt
requests>=2.31.0
pydub>=0.25.1
google-cloud-texttospeech>=2.14.0
python-dotenv>=1.0.0
lameenc>=1.7.0
